second_task/requirements.txt
torch
torchvision
transformers
datasets
scikit-learn
pillow
numpy
pandas
matplotlib
tqdm
jupyter
ipython